#13625: Update golden function of lerp

+++ ttnn/ttnn/operations/ternary.py
@@ -37,7 +37,7 @@ ttnn.attach_golden_function(ttnn.addcdiv, golden_function=_golden_function_addcd
 def _golden_function_lerp(input_tensor_a, input_tensor_b, input_tensor_c, *args, **kwargs):
     import torch
 
-    return torch.lerp(input_tensor_a, input_tensor_b, input_tensor_c)
+    return torch.lerp(input_tensor_a, input_tensor_b.to(input_tensor_a.dtype), input_tensor_c.to(input_tensor_a.dtype))
 
 
 ttnn.attach_golden_function(ttnn.lerp, golden_function=_golden_function_lerp)
